Refactor parsing logic
