+++ basyx.aasenvironment/basyx.aasenvironment-client/src/test/java/org/eclipse/digitaltwin/basyx/aasenvironment/client/resolvers/RegistryDescriptorResolverTest.java
@@ -31,8 +31,6 @@ import org.eclipse.digitaltwin.aas4j.v3.model.AssetAdministrationShell;
 import org.eclipse.digitaltwin.aas4j.v3.model.Submodel;
 import org.eclipse.digitaltwin.basyx.aasenvironment.client.DummyAasEnvironmentComponent;
 import org.eclipse.digitaltwin.basyx.aasenvironment.client.TestFixture;
-import org.eclipse.digitaltwin.basyx.aasenvironment.client.resolvers.parsers.AasRegistryEndpointURIParser;
-import org.eclipse.digitaltwin.basyx.aasenvironment.client.resolvers.parsers.SubmodelRegistryEndpointURIParser;
 import org.eclipse.digitaltwin.basyx.aasregistry.client.ApiException;
 import org.eclipse.digitaltwin.basyx.aasrepository.AasRepository;
 import org.eclipse.digitaltwin.basyx.submodelrepository.SubmodelRepository;
@@ -76,7 +74,7 @@ public class RegistryDescriptorResolverTest {
 
 	@Test
 	public void resolveAasDescriptor() throws ApiException {
-		AasDescriptorResolver resolver = new AasDescriptorResolver(new EndpointResolver<>(new AasRegistryEndpointURIParser()));
+		AasDescriptorResolver resolver = new AasDescriptorResolver(new EndpointResolver());
 		
 		AssetAdministrationShell expectedAas = FIXTURE.buildAasPre1();
 		AssetAdministrationShell actualAas = resolver.resolveAasDescriptor(FIXTURE.buildAasPre1Descriptor()).getAAS();
@@ -86,7 +84,7 @@ public class RegistryDescriptorResolverTest {
 
 	@Test
 	public void resolveSmDescriptor() throws ApiException {
-		SubmodelDescriptorResolver resolver = new SubmodelDescriptorResolver(new EndpointResolver<>(new SubmodelRegistryEndpointURIParser()));
+		SubmodelDescriptorResolver resolver = new SubmodelDescriptorResolver(new EndpointResolver());
 
 		Submodel expectedSm = FIXTURE.buildSmPre1();
 		Submodel actualSm = resolver.resolveSubmodelDescriptor(FIXTURE.buildSmPre1Descriptor()).getSubmodel();
